# Conflicts:
#	config/scripts/00_ddl-schema.sql

+++ src/main/java/com/gol/ants_quests/hibernate/entities/DomandaQuest.java
@@ -28,7 +28,7 @@ public class DomandaQuest extends GenericEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idQstDet;
+    private Long idQstDet;
     private String domanda;
 
     @JsonBackReference
