+++ src/main/java/com/gol/ants_quests/hibernate/services/RisposteHibService.java
@@ -6,7 +6,7 @@ import com.gol.ants_quests.hibernate.entities.RispostaQuest;
 import com.gol.ants_quests.hibernate.repositories.RisposteRepository;
 
 @Service
-public class RisposteHibService extends GenericHibService<RispostaQuest, Integer, RisposteRepository> {
+public class RisposteHibService extends GenericHibService<RispostaQuest, Long, RisposteRepository> {
 
     public RisposteHibService(RisposteRepository repository) {
         super(repository);
